+++ Devtools/KeyConsole.cs
@@ -180,7 +180,7 @@ namespace Sideload.Devtools
     {
         private static bool Prefix(string args)
         {
-            try { return !KeyConsole.TryHandle(args) && !SurfaceConsole.TryHandle(args); } catch { return true; }
+            try { return !KeyConsole.TryHandle(args) && !SurfaceConsole.TryHandle(args) && !WheelConsole.TryHandle(args); } catch { return true; }
         }
     }
 
@@ -190,7 +190,7 @@ namespace Sideload.Devtools
     {
         private static bool Prefix(Il2CppSystem.Collections.Generic.List<string> args)
         {
-            try { return !KeyConsole.TryHandle(args) && !SurfaceConsole.TryHandle(args); } catch { return true; }
+            try { return !KeyConsole.TryHandle(args) && !SurfaceConsole.TryHandle(args) && !WheelConsole.TryHandle(args); } catch { return true; }
         }
     }
 }